fix: java compiler errors
