+++ src/objs/AxeType.java
@@ -1,19 +1,6 @@
 package objs;
 
-public static class AxeTypeInput {
-	public static final int WOODEN = 1;
-	public static final int STONE = 2;
-	public static final int IRON = 3;
-	public static final int GOLDEN = 4;
-	public static final int DIAMOND = 5;
-	public static final int NETHERITE = 6;
-
-	public boolean isValid(int value) {
-		return value >= AxeTypeInput.WOODEN && value <= AxeTypeInput.NETHERITE;
-	}
-}
-
-public static class AxeType {
+public class AxeType {
 	public static final String WOODEN = "WOODEN_AXE";
 	public static final String STONE = "STONE_AXE";
 	public static final String IRON = "IRON_AXE";
